Make `tctl get` always fetch by name when name is specified
For some resource kinds, `tctl get` used to always return the list of
all resources of that kind, irrespective of whether a specific resource
name was given in the argument to `tctl get`.

For instance, `tctl get node/node-id` would have always listed all the
nodes (just like mere `tctl get node`).

In those cases when the underlying API does not provide a method to fetch
resources by name, `tctl get` is hereby adapted to return an error:
```
ERROR: getting "node/node-id" is not supported
```
(NB: This is the response even when `node-id` does actually exist.)

In the other cases when the behaviour was just due to an omission in the
code (e.g., SSO connectors) the relevant `tctl get` handlers are fixed.